+++ week4/week4assignment/src/components/Router.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-
-import Header from "./Header.jsx";
-import Content from "./Content.jsx";
+import React from "react";
+import Header from "./Header.tsx";
+import Content from "./Content.tsx";
 
 function Router() {
   return (
